# docs/swagger.yaml
basePath: /v1
definitions:
  domain.KPIEffect:
    properties:
      direction:
        description: increase | decrease
        type: string
      magnitude:
        type: string
      metric:
        type: string
    type: object
  domain.RankingWeights:
    properties:
      evidence:
        type: number
      feasibility:
        type: number
      impact:
        type: number
      novelty:
        type: number
      risk:
        type: number
    type: object
  domain.Scores:
    properties:
      confidence:
        type: number
      evidenceStrength:
        type: number
      feasibility:
        type: number
      impact:
        type: number
      novelty:
        type: number
      riskPenalty:
        type: number
      total:
        type: number
    type: object
  domain.VerificationStep:
    properties:
      estimatedCost:
        type: string
      estimatedDuration:
        type: string
      resource:
        type: string
      step:
        type: string
      successCriterion:
        type: string
    type: object
  errs.Error:
    properties:
      caller:
        type: string
      details:
        items:
          $ref: '#/definitions/errs.FieldError'
        type: array
      message:
        type: string
      reason:
        type: string
      type:
        $ref: '#/definitions/errs.Type'
    type: object
  errs.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
  errs.Type:
    enum:
    - VALIDATION
    - NOT_FOUND
    - CONFLICT
    - TIMEOUT
    - INTERNAL
    type: string
    x-enum-comments:
      ErrTypeConflict: "409"
      ErrTypeInternal: "500"
      ErrTypeNotFound: "404"
      ErrTypeTimeout: "504"
      ErrTypeValidation: "422"
    x-enum-descriptions:
    - "422"
    - "404"
    - "409"
    - "504"
    - "500"
    x-enum-varnames:
    - ErrTypeValidation
    - ErrTypeNotFound
    - ErrTypeConflict
    - ErrTypeTimeout
    - ErrTypeInternal
  in.CreateRunRequest:
    properties:
      domain:
        description: Domain — база знаний какого домена используется (по умолчанию
          flotation).
        example: flotation
        type: string
      excludedTopics:
        description: |-
          ExcludedTopics — направления/подходы, которые генерация должна
          обходить (например, уже испробованные и отклонённые ранее).
        items:
          type: string
        type: array
      language:
        description: Language — язык гипотез/обоснований в ответе (по умолчанию ru).
        enum:
        - ru
        - en
        - zh
        example: ru
        type: string
      rankingWeights:
        allOf:
        - $ref: '#/definitions/domain.RankingWeights'
        description: |-
          RankingWeights — режим экспертной настройки: переопределение весов
          критериев ранжирования (нетронутые поля = дефолт).
      rawInput:
        additionalProperties: {}
        type: object
      rawText:
        example: 'Фабрика: КГМК. Породные хвосты 5824591 СМТ...'
        type: string
    required:
    - rawText
    type: object
...
  out.GraphEdgeResponse:
    properties:
      from:
        type: string
      id:
        type: string
      label:
        type: string
      to:
        type: string
      type:
        description: subject | affects | evidence
        example: evidence
        type: string
    type: object
  out.GraphNodeResponse:
    properties:
      id:
        type: string
      kind:
        type: string
      label:
        type: string
      type:
        description: entity | claim | hypothesis
        example: entity
        type: string
    type: object
  out.GraphResponse:
    properties:
      edges:
        items:
          $ref: '#/definitions/out.GraphEdgeResponse'
        type: array
      nodes:
        items:
          $ref: '#/definitions/out.GraphNodeResponse'
        type: array
    type: object
  out.HypothesisResponse:
    properties:
      createdAt:
        type: string
      criticNotes:
        type: string
      evidenceRefs:
        items:
          type: string
        type: array
      expectedKpiEffect:
        $ref: '#/definitions/domain.KPIEffect'
      id:
        type: string
      mechanism:
        type: string
      noveltyReason:
        type: string
      rank:
        type: integer
      risks:
        items:
          type: string
        type: array
      runId:
        type: string
      scores:
        $ref: '#/definitions/domain.Scores'
      statement:
        type: string
      verificationPlan:
        items:
          $ref: '#/definitions/domain.VerificationStep'
        type: array
    type: object
...
  out.ProblemSpecResponse:
    properties:
      availableEquipment:
        items:
          type: string
        type: array
      constraints:
        items:
          type: string
        type: array
      horizon:
        type: string
      lossHotspots:
        items:
          type: string
        type: array
      plant:
        type: string
      targetKpi:
        type: string
      targetMetals:
        items:
          type: string
        type: array
    type: object
  out.RunListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/out.RunResponse'
        type: array
      page:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
  out.RunResponse:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      domain:
        type: string
      error:
        type: string
      hypotheses:
        items:
          $ref: '#/definitions/out.HypothesisResponse'
        type: array
      id:
        type: string
      knowledgeGaps:
        description: |-
          KnowledgeGaps — металлы/точки потерь из ProblemSpec, слабо покрытые
          извлечёнными claims (детерминированная проверка, не LLM-догадка).
        items:
          type: string
        type: array
      language:
        type: string
      problemSpec:
        $ref: '#/definitions/out.ProblemSpecResponse'
      status:
        example: retrieving
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'Evidence-Backed Hypothesis Factory: генерация проверяемых технологических
    гипотез по обогащению руд на основе базы знаний (RAG).'
  title: Фабрика гипотез — Hypothesis Factory API
  version: "1.0"
paths:
...
  /runs:
    get:
      parameters:
      - default: 1
        description: Номер страницы (>=1)
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы (1-100)
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/out.RunListResponse'
      summary: Список прогонов
      tags:
      - runs
    post:
      consumes:
      - application/json
      description: Принимает свободный текст (профиль хвостов/проблему), возвращает
        run сразу в статусе retrieving — дальше поллить GET /runs/{runId}.
      parameters:
      - description: Запрос на генерацию
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/in.CreateRunRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/out.RunResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Запуск генерации гипотез
      tags:
      - runs
  /runs/{runId}:
    get:
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/out.RunResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Статус и результат прогона
      tags:
      - runs
  /runs/{runId}/graph:
    get:
      description: 'Узлы: entity (оборудование/показатель/реагент/...), claim, hypothesis.
        Рёбра: subject (entity→claim), affects (claim→entity), evidence (claim→hypothesis).'
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/out.GraphResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Граф evidence-связей прогона
      tags:
      - runs
  /runs/{runId}/report.csv:
    get:
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: CSV-отчёт по прогону
      tags:
      - runs
  /runs/{runId}/report.docx:
    get:
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.wordprocessingml.document
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: DOCX-отчёт по прогону
      tags:
      - runs
  /runs/{runId}/report.jira.json:
    get:
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      - description: Ключ проекта Jira, напр. HYP
        in: query
        name: projectKey
        required: true
        type: string
      - default: Task
        description: Тип задачи
        in: query
        name: issueType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Экспорт гипотез как задач в Jira-совместимом JSON
      tags:
      - runs
  /runs/{runId}/report.md:
    get:
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      produces:
      - text/markdown
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Markdown-отчёт по прогону
      tags:
      - runs
  /runs/{runId}/report.pdf:
    get:
      parameters:
      - description: UUID прогона
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: PDF-отчёт по прогону
      tags:
      - runs
  /runs/from-excel:
    post:
      consumes:
      - multipart/form-data
      description: Принимает Хвосты *.xlsx — числа (классы крупности, минеральные
        формы, Ni/Cu) парсятся детерминированно, без LLM. rawText опционален — дополняет
        качественные поля (цель, оборудование, ограничения).
      parameters:
      - description: Хвосты *.xlsx
        in: formData
        name: file
        required: true
        type: file
      - description: Свободный текст с доп. контекстом (цель, оборудование, ограничения)
        in: formData
        name: rawText
        type: string
      - default: flotation
        description: База знаний какого домена используется
        in: formData
        name: domain
        type: string
      - default: ru
        description: 'Язык гипотез: ru|en|zh'
        in: formData
        name: language
        type: string
      - description: JSON domain.RankingWeights — режим экспертной настройки весов
        in: formData
        name: rankingWeights
        type: string
      - description: JSON-массив строк — направления, которые генерация должна обходить
        in: formData
        name: excludedTopics
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/out.RunResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Запуск генерации гипотез из профиля хвостов (Excel)
      tags:
      - runs
swagger: "2.0"