# in/run.go
package in

import "hypothesis-factory/domain"

type CreateRunRequest struct {
	RawText  string         `json:"rawText"  validate:"required" example:"Фабрика: КГМК. Породные хвосты 5824591 СМТ..."`
	RawInput map[string]any `json:"rawInput"`
	// Domain — база знаний какого домена используется (по умолчанию flotation).
	Domain string `json:"domain" example:"flotation"`
	// Language — язык гипотез/обоснований в ответе (по умолчанию ru).
	Language string `json:"language" validate:"omitempty,oneof=ru en zh" example:"ru"`
	// RankingWeights — режим экспертной настройки: переопределение весов
	// критериев ранжирования (нетронутые поля = дефолт).
	RankingWeights *domain.RankingWeights `json:"rankingWeights"`
	// ExcludedTopics — направления/подходы, которые генерация должна
	// обходить (например, уже испробованные и отклонённые ранее).
	ExcludedTopics []string `json:"excludedTopics"`
}